Update telemetry data types from `int` to `int64` (#5185)

+++ internal/telemetry/cluster.go
@@ -8,12 +8,12 @@ import (
 
 // NodeCount returns the total number of nodes in the cluster.
 // It returns an error if the underlying k8s API client errors.
-func (c *Collector) NodeCount(ctx context.Context) (int, error) {
+func (c *Collector) NodeCount(ctx context.Context) (int64, error) {
 	nodes, err := c.Config.K8sClientReader.CoreV1().Nodes().List(ctx, metaV1.ListOptions{})
 	if err != nil {
 		return 0, err
 	}
-	return len(nodes.Items), nil
+	return int64(len(nodes.Items)), nil
 }
 
 // ClusterID returns the UID of the kube-system namespace representing cluster id.
